requirements.txt
PyYAML == 5.3.1
pytest == 6.2.1
jsonSchema == 3.2.0
requests == 2.25.1